requirements.txt
binance-futures-connector
python-telegram-bot
pandas
yagmail
websockets
google-generativeai
newsapi-python
fastapi
bybit-ws
pyodbc
websocket-client
